Create extensions module
